Refs #35078 - move redux to SystemPurposeCard
